Delegate environment lookup to Atom#evaluate
Evaluate the first item in a list so that the environment lookup is in a
single place (on Atom). This has the pleasant side effect of allowing
us to evaluate lists who's first element is not an Atom.

The test for "lambda" in Chapter 2 now passes, so we remove the
'pending' flag.